+++ NEWS.md
@@ -4,6 +4,7 @@
 
 * Include an option to turn off Y axis inversion in pipeline functions.  
 * Add support for gzip compressed transcript data file. 
+* Expand version compatibility to igraph 2.1.0 and above.
 
 # FastReseg 1.0.2